Weird First Implementation of Boss
Boss is sort of in. Attack timing is still in bug testing, and it'll attack even when the player is out of range at the moment. Only 1 attack is in (line attack). Doesn't draw with the correct size (hit box is 3 times larger than image). But hey, it's a start ¯\_(ツ)_/¯.